common: changed long double to double

+++ src/common/Node.h
@@ -316,10 +316,10 @@ namespace common {
 
 	class Float : public Pattern {
 	public:
-		long double value;
+		double value;
 
-		Float(long double);
-		Float(long double, int);
+		Float(double);
+		Float(double, int);
 
 		virtual void accept(Visitor &);
 	};
